feat: Updates for use in worksheet
The machine learning usage here was inverted - PLAYER and OPPONENT
is not the same as PLAYER and COMPUTER. If you train a machine
learning model with moves made by the human, and then try and
make predictions with the board described in the same way, then
the machine learning model is being trained how to lose.

I've separated out HUMAN/COMPUTER and PLAYER/OPPONENT which means
the machine learning model can be trained from winning games
by either player.

I've also done some refactoring and tidying:
- Adding comments to explain the code so it is more useful as
  an educational project
- Using variables and constants instead of copy/pasted code
- Using variables and constants instead of magic numbers

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>